Refactor Completions logic to be centralized to CompletionsProcessor
