+++ src/files/BrsFile.spec.ts
@@ -1,6 +1,6 @@
 import { assert, expect } from '../chai-config.spec';
 import * as sinonImport from 'sinon';
-import { CompletionItemKind, Position, Range } from 'vscode-languageserver';
+import { Position, Range } from 'vscode-languageserver';
 import type { Callable, CommentFlag, VariableDeclaration } from '../interfaces';
 import { Program } from '../Program';
 import { BooleanType } from '../types/BooleanType';
@@ -11,12 +11,11 @@ import { StringType } from '../types/StringType';
 import { BrsFile } from './BrsFile';
 import { SourceMapConsumer } from 'source-map';
 import { Lexer } from '../lexer/Lexer';
-import { TokenKind, Keywords } from '../lexer/TokenKind';
+import { TokenKind } from '../lexer/TokenKind';
 import { DiagnosticMessages } from '../DiagnosticMessages';
-import type { StandardizedFileEntry } from 'roku-deploy';
 import util, { standardizePath as s } from '../util';
 import PluginInterface from '../PluginInterface';
-import { expectCompletionsIncludes, expectDiagnostics, expectHasDiagnostics, expectTypeToBe, expectZeroDiagnostics, getTestGetTypedef, getTestTranspile, trim } from '../testHelpers.spec';
+import { expectDiagnostics, expectHasDiagnostics, expectTypeToBe, expectZeroDiagnostics, getTestGetTypedef, getTestTranspile, trim } from '../testHelpers.spec';
 import { ParseMode } from '../parser/Parser';
 import { Logger } from '../Logger';
 import { ImportStatement } from '../parser/Statement';
@@ -172,290 +171,11 @@ describe('BrsFile', () => {
         expectDiagnostics(file, expected);
     });
 
-    describe('getPartialVariableName', () => {
-        let entry = {
-            src: `${rootDir}/source/lib.brs`,
-            dest: `source/lib.brs`
-        } as StandardizedFileEntry;
-
-        it('creates proper tokens', () => {
-            file = program.setFile<BrsFile>(entry, `call(ModuleA.ModuleB.ModuleC.`);
-            expect(file['getPartialVariableName'](file.parser.tokens[7])).to.equal('ModuleA.ModuleB.ModuleC.');
-            expect(file['getPartialVariableName'](file.parser.tokens[6])).to.equal('ModuleA.ModuleB.ModuleC');
-            expect(file['getPartialVariableName'](file.parser.tokens[5])).to.equal('ModuleA.ModuleB.');
-            expect(file['getPartialVariableName'](file.parser.tokens[4])).to.equal('ModuleA.ModuleB');
-            expect(file['getPartialVariableName'](file.parser.tokens[3])).to.equal('ModuleA.');
-            expect(file['getPartialVariableName'](file.parser.tokens[2])).to.equal('ModuleA');
-        });
-    });
-
     describe('getScopesForFile', () => {
         it('finds the scope for the file', () => {
             let file = program.setFile('source/main.brs', ``);
             expect(program.getScopesForFile(file)[0]?.name).to.equal('source');
         });
-    });
-
-    describe('getCompletions', () => {
-        it('does not crash for callfunc on a function call', () => {
-            const file = program.setFile('source/main.brs', `
-                sub main()
-                    getManager()@.
-                end sub
-            `);
-            expect(() => {
-                program.getCompletions(file.srcPath, util.createPosition(2, 34));
-            }).not.to.throw;
-        });
-
-        it('suggests pkg paths in strings that match that criteria', () => {
-            program.setFile('source/main.brs', `
-                sub main()
-                    print "pkg:"
-                end sub
-            `);
-            const result = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(2, 31));
-            const names = result.map(x => x.label);
-            expect(names.sort()).to.eql([
-                'pkg:/source/main.brs'
-            ]);
-        });
-
-        it('suggests libpkg paths in strings that match that criteria', () => {
-            program.setFile('source/main.brs', `
-                sub main()
-                    print "libpkg:"
-                end sub
-            `);
-            const result = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(2, 31));
-            const names = result.map(x => x.label);
-            expect(names.sort()).to.eql([
-                'libpkg:/source/main.brs'
-            ]);
-        });
-
-        it('suggests pkg paths in template strings', () => {
-            program.setFile('source/main.brs', `
-                sub main()
-                    print \`pkg:\`
-                end sub
-            `);
-            const result = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(2, 31));
-            const names = result.map(x => x.label);
-            expect(names.sort()).to.eql([
-                'pkg:/source/main.brs'
-            ]);
-        });
-
-        it('waits for the file to be processed before collecting completions', () => {
-            //eslint-disable-next-line @typescript-eslint/no-floating-promises
-            program.setFile('source/main.brs', `
-                sub Main()
-                    print "hello"
-                    Say
-                end sub
-
-                sub SayHello()
-                end sub
-            `);
-
-            let result = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(3, 23));
-            let names = result.map(x => x.label);
-            expect(names).to.includes('Main');
-            expect(names).to.includes('SayHello');
-        });
-
-        it('includes every type of item at base level', () => {
-            program.setFile('source/main.bs', `
-                sub main()
-                    print
-                end sub
-                sub speak()
-                end sub
-                namespace stuff
-                end namespace
-                class Person
-                end class
-                enum Direction
-                end enum
-            `);
-            expectCompletionsIncludes(program.getCompletions('source/main.bs', util.createPosition(2, 26)), [{
-                label: 'main',
-                kind: CompletionItemKind.Function
-            }, {
-                label: 'speak',
-                kind: CompletionItemKind.Function
-            }, {
-                label: 'stuff',
-                kind: CompletionItemKind.Module
-            }, {
-                label: 'Person',
-                kind: CompletionItemKind.Class
-            }, {
-                label: 'Direction',
-                kind: CompletionItemKind.Enum
-            }]);
-        });
-
-        describe('namespaces', () => {
-            it('gets full namespace completions at any point through the leading identifier', () => {
-                program.setFile('source/main.bs', `
-                    sub main()
-                        foo.bar
-                    end sub
-
-                    namespace foo.bar
-                    end namespace
-
-                    class Person
-                    end class
-                `);
-
-                const result = program.getCompletions(`${rootDir}/source/main.bs`, Position.create(2, 24)).map(x => x.label);
-                expect(result).includes('main');
-                expect(result).includes('foo');
-                expect(result).includes('Person');
-            });
-
-            it('gets namespace completions', () => {
-                program.setFile('source/main.bs', `
-                    namespace foo.bar
-                        function sayHello()
-                        end function
-                    end namespace
-
-                    sub Main()
-                        print "hello"
-                        foo.ba
-                        foo.bar.
-                    end sub
-                `);
-
-                let result = program.getCompletions(`${rootDir}/source/main.bs`, Position.create(8, 30));
-                let names = result.map(x => x.label);
-                expect(names).to.includes('bar');
-
-                result = program.getCompletions(`${rootDir}/source/main.bs`, Position.create(9, 32));
-                names = result.map(x => x.label);
-                expect(names).to.includes('sayHello');
-            });
-        });
-
-        it('always includes `m`', () => {
-            //eslint-disable-next-line @typescript-eslint/no-floating-promises
-            program.setFile('source/main.brs', `
-                sub Main()
-
-                end sub
-            `);
-
-            let result = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(2, 23));
-            let names = result.map(x => x.label);
-            expect(names).to.contain('m');
-        });
-
-        it('does not fail for missing previousToken', () => {
-            //add a single character to the file, and get completions after it
-            program.setFile('source/main.brs', `i`);
-            expect(() => {
-                program.getCompletions(`${rootDir}/source/main.brs`, Position.create(0, 1)).map(x => x.label);
-            }).not.to.throw;
-        });
-
-        it('includes all keywords`', () => {
-            //eslint-disable-next-line @typescript-eslint/no-floating-promises
-            program.setFile('source/main.brs', `
-                sub Main()
-
-                end sub
-            `);
-
-            let keywords = Object.keys(Keywords).filter(x => !x.includes(' '));
-
-            //inside the function
-            let result = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(2, 23));
-            let names = result.map(x => x.label);
-            for (let keyword of keywords) {
-                expect(names).to.include(keyword);
-            }
-
-            //outside the function
-            result = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(4, 8));
-            names = result.map(x => x.label);
-            for (let keyword of keywords) {
-                expect(names).to.include(keyword);
-            }
-        });
-
-        it('does not provide completions within a comment', () => {
-            //eslint-disable-next-line @typescript-eslint/no-floating-promises
-            program.setFile('source/main.brs', `
-                sub Main()
-                    'some comment
-                end sub
-            `);
-
-            //inside the function
-            let result = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(2, 33));
-            expect(result).to.be.lengthOf(0);
-        });
-
-        it('does not provide duplicate entries for variables', () => {
-            //eslint-disable-next-line @typescript-eslint/no-floating-promises
-            program.setFile('source/main.brs', `
-                sub Main()
-                    name = "bob"
-                    age = 12
-                    name = "john"
-                end sub
-            `);
-
-            let result = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(3, 23));
-
-            let count = result.reduce((total, x) => {
-                return x.label === 'name' ? total + 1 : total;
-            }, 0);
-            expect(count).to.equal(1);
-        });
-
-        it('does not include `as` and `string` text options when used in function params', () => {
-            //eslint-disable-next-line @typescript-eslint/no-floating-promises
-            program.setFile('source/main.brs', `
-                sub Main(name as string)
-
-                end sub
-            `);
-
-            let result = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(2, 23));
-            expect(result.filter(x => x.kind === CompletionItemKind.Text)).not.to.contain('as');
-            expect(result.filter(x => x.kind === CompletionItemKind.Text)).not.to.contain('string');
-        });
-
-        it('does not provide intellisense results when inside a comment', () => {
-            //eslint-disable-next-line @typescript-eslint/no-floating-promises
-            program.setFile('source/main.brs', `
-                sub Main(name as string)
-                    'this is a comment
-                end sub
-            `);
-
-            let results = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(2, 30));
-            expect(results).to.be.empty;
-        });
-
-        it('does provide intellisence for labels only after a goto keyword', () => {
-            //eslint-disable-next-line @typescript-eslint/no-floating-promises
-            program.setFile('source/main.brs', `
-                sub Main(name as string)
-                    something:
-                    goto \nend sub
-            `);
-
-            let results = program.getCompletions(`${rootDir}/source/main.brs`, Position.create(3, 25));
-            expect(results.length).to.equal(1);
-            expect(results[0]?.label).to.equal('something');
-        });
-
     });
 
     describe('comment flags', () => {
